You can't always use %ld for off_t
Depending on your operating system, off_t may be 32 bits (%ld) or may be
64 bits (%lld).  It's easiest to cast it to a (long long) and just use
%lld in all cases.

See the discussion here:

  http://stackoverflow.com/questions/586928/how-should-i-print-types-like-off-t-and-size-t

+++ fts.c
@@ -843,8 +843,9 @@ _Bool WriteMailContent(FILE *File)
               snprintf(OutBuffer, DEFAULT_BUFFER_SIZE - 1,
                 "  - %s (%sytes)\r\n", Help, TempBuffer);
             else
+              /* lks: http://stackoverflow.com/questions/586928/how-should-i-print-types-like-off-t-and-size-t */
               snprintf(OutBuffer, DEFAULT_BUFFER_SIZE - 1,
-                "  - %s (%ld Bytes)\r\n", Help, Response->Size);
+                "  - %s (%lld Bytes)\r\n", Help, (long long)Response->Size);
 
             fputs(OutBuffer, File);
           }
